1. (2.) Poprawiono bledy zwiazane z formatowaniem kodu
