+++ src/main/java/edu/kis/vh/nursery/factory/Rhymersfactory.java
@@ -1,15 +1,14 @@
 package edu.kis.vh.nursery.factory;
 
-import edu.kis.vh.nursery.defaultCountingOutRhymer;
+import edu.kis.vh.nursery.DefaultCountingOutRhymer;
 
-	public interface Rhymersfactory {
-	
-		public defaultCountingOutRhymer GetStandardRhymer();
-		
-		public defaultCountingOutRhymer GetFalseRhymer();
-		
-		public defaultCountingOutRhymer GetFIFORhymer();
-		
-		public defaultCountingOutRhymer GetHanoiRhymer();
-		
-	}
+public interface Rhymersfactory {
+
+    DefaultCountingOutRhymer GetStandardRhymer();
+
+    DefaultCountingOutRhymer GetFalseRhymer();
+
+    DefaultCountingOutRhymer GetFIFORhymer();
+
+    DefaultCountingOutRhymer GetHanoiRhymer();
+}
